--- a/graph/traverse/traverse.go
+++ b/graph/traverse/traverse.go
@@ -6,10 +6,9 @@
 package traverse // import "gonum.org/v1/gonum/graph/traverse"
 
 import (
-	"golang.org/x/tools/container/intsets"
-
 	"gonum.org/v1/gonum/graph"
 	"gonum.org/v1/gonum/graph/internal/linear"
+	"gonum.org/v1/gonum/graph/internal/set"
 )
 
 // BreadthFirst implements stateful breadth-first graph traversal.
@@ -17,7 +16,7 @@ type BreadthFirst struct {
 	EdgeFilter func(graph.Edge) bool
 	Visit      func(u, v graph.Node)
 	queue      linear.NodeQueue
-	visited    *intsets.Sparse
+	visited    set.Ints
 }
 
 // Walk performs a breadth-first traversal of the graph g starting from the given node,
@@ -27,10 +26,10 @@ type BreadthFirst struct {
 // non-nil, it is called with the nodes joined by each followed edge.
 func (b *BreadthFirst) Walk(g graph.Graph, from graph.Node, until func(n graph.Node, d int) bool) graph.Node {
 	if b.visited == nil {
-		b.visited = &intsets.Sparse{}
+		b.visited = make(set.Ints)
 	}
 	b.queue.Enqueue(from)
-	b.visited.Insert(from.ID())
+	b.visited.Add(from.ID())
 
 	var (
 		depth     int
@@ -52,7 +51,7 @@ func (b *BreadthFirst) Walk(g graph.Graph, from graph.Node, until func(n graph.N
 			if b.Visit != nil {
 				b.Visit(t, n)
 			}
-			b.visited.Insert(n.ID())
+			b.visited.Add(n.ID())
 			children++
 			b.queue.Enqueue(n)
 		}
@@ -93,15 +92,13 @@ func (b *BreadthFirst) WalkAll(g graph.Undirected, before, after func(), during 
 
 // Visited returned whether the node n was visited during a traverse.
 func (b *BreadthFirst) Visited(n graph.Node) bool {
-	return b.visited != nil && b.visited.Has(n.ID())
+	return b.visited.Has(n.ID())
 }
 
 // Reset resets the state of the traverser for reuse.
 func (b *BreadthFirst) Reset() {
 	b.queue.Reset()
-	if b.visited != nil {
-		b.visited.Clear()
-	}
+	b.visited = nil
 }
 
 // DepthFirst implements stateful depth-first graph traversal.
@@ -109,7 +106,7 @@ type DepthFirst struct {
 	EdgeFilter func(graph.Edge) bool
 	Visit      func(u, v graph.Node)
 	stack      linear.NodeStack
-	visited    *intsets.Sparse
+	visited    set.Ints
 }
 
 // Walk performs a depth-first traversal of the graph g starting from the given node,
@@ -119,10 +116,10 @@ type DepthFirst struct {
 // is called with the nodes joined by each followed edge.
 func (d *DepthFirst) Walk(g graph.Graph, from graph.Node, until func(graph.Node) bool) graph.Node {
 	if d.visited == nil {
-		d.visited = &intsets.Sparse{}
+		d.visited = make(set.Ints)
 	}
 	d.stack.Push(from)
-	d.visited.Insert(from.ID())
+	d.visited.Add(from.ID())
 
 	for d.stack.Len() > 0 {
 		t := d.stack.Pop()
@@ -139,7 +136,7 @@ func (d *DepthFirst) Walk(g graph.Graph, from graph.Node, until func(graph.Node)
 			if d.Visit != nil {
 				d.Visit(t, n)
 			}
-			d.visited.Insert(n.ID())
+			d.visited.Add(n.ID())
 			d.stack.Push(n)
 		}
 	}
@@ -174,13 +171,11 @@ func (d *DepthFirst) WalkAll(g graph.Undirected, before, after func(), during fu
 
 // Visited returned whether the node n was visited during a traverse.
 func (d *DepthFirst) Visited(n graph.Node) bool {
-	return d.visited != nil && d.visited.Has(n.ID())
+	return d.visited.Has(n.ID())
 }
 
 // Reset resets the state of the traverser for reuse.
 func (d *DepthFirst) Reset() {
 	d.stack = d.stack[:0]
-	if d.visited != nil {
-		d.visited.Clear()
-	}
+	d.visited = nil
 }