--- a/graph/undirect.go
+++ b/graph/undirect.go
@@ -4,10 +4,6 @@
 
 package graph
 
-import (
-	"golang.org/x/tools/container/intsets"
-)
-
 // Undirect converts a directed graph to an undirected graph, resolving
 // edge weight conflicts.
 type Undirect struct {
@@ -47,20 +43,18 @@ func (g Undirect) Nodes() []Node { return g.G.Nodes() }
 
 // From returns all nodes in g that can be reached directly from u.
 func (g Undirect) From(u Node) []Node {
-	var (
-		nodes []Node
-		seen  intsets.Sparse
-	)
+	var nodes []Node
+	seen := make(map[int]struct{})
 	for _, n := range g.G.From(u) {
-		seen.Insert(n.ID())
+		seen[n.ID()] = struct{}{}
 		nodes = append(nodes, n)
 	}
 	for _, n := range g.G.To(u) {
 		id := n.ID()
-		if seen.Has(id) {
+		if _, ok := seen[id]; ok {
 			continue
 		}
-		seen.Insert(id)
+		seen[n.ID()] = struct{}{}
 		nodes = append(nodes, n)
 	}
 	return nodes