… graph/community/louvain_undirected_multiplex_test.go
… graph/community/louvain_undirected_test.go
… graph/encoding/dot/dot_test.go