This is a first step towards allowing `int64` ID values and addresses the simple-to-handle cases.

Another PR for the more hairy community package will follow when I figure out a good approach; the problem is that `intsets.Sparse` iterates via a stateful method, which allows a nice method-based approach to handle subset iteration, versus complete set iteration without filling a complete set for the complete set case. `set.Ints` cannot do this because of limitations with `range`. Suggestions welcomed.

@vladimir-ch @mewmew Please take a look.