@kortschak 2017-06-15
The approach to maintaining an iterator required for the multiplex cases is handled by a `minTaker` now in `louvain_common.go`. This also works in a cut down manner for the non-multiplex case.

The impact on performance is not terrible:

```
$ benchstat old.bench nosort.bench
name                        old time/op  new time/op  delta
LouvainDirectedMultiplex-8  27.6ms ± 2%  29.4ms ± 2%  +6.58%  (p=0.000 n=10+10)
LouvainDirected-8           24.2ms ± 3%  24.7ms ± 2%  +2.07%  (p=0.022 n=10+9)
LouvainMultiplex-8          21.1ms ± 2%  22.3ms ± 4%  +5.79%  (p=0.000 n=10+10)
Louvain-8                   18.9ms ± 2%  19.3ms ± 2%  +2.33%  (p=0.000 n=10+10)
```
except for the fact that without consistent sort order of the nodes (which `intsets.Sparse` provided) we do not pass the tests. I think that it is worth being able to control the source of randomness in this package for reproducibility. So the IDs are sorted giving a worse performance (the code that is committed here).
```
$ benchstat old.bench new.bench
name                        old time/op  new time/op  delta
LouvainDirectedMultiplex-8  27.6ms ± 2%  30.8ms ± 3%  +11.60%  (p=0.000 n=10+10)
LouvainDirected-8           24.2ms ± 3%  26.5ms ± 3%   +9.50%  (p=0.000 n=10+10)
LouvainMultiplex-8          21.1ms ± 2%  24.1ms ± 4%  +14.43%  (p=0.000 n=10+10)
Louvain-8                   18.9ms ± 2%  20.7ms ± 3%   +9.74%  (p=0.000 n=10+10)
```

This can be fixed later.