@mewmew 2017-06-16
From what I can tell, the change looks good to me.

When reviewing the change, and especially as you commented on the performance impact of changing from `intsets.Sparse` to `map[int]struct{}`, I feel curious to ask if it would make sense to fork https://golang.org/x/tools/container/intsets and maintain a `int64` version of it?

That could make the chance less intrusive and still maintain a good performance. What are your thoughts?