@kortschak 2017-06-16
I had a look at the code for intsets and it's not trivial to make it a `int64`-handling package, and that adds significant maintenance burden.