@kortschak 2017-06-16
What you will notice though is that in the bigids branch, the parts of community that handle this deal with ids that are in `int`, so after that has gone it, I can revert the parts of this PR this have caused the slow down. It's a bit of a rigmarole, but it the change involved in some of the routines from `int` -> `int64` are not as simple as just a mechanical change. This is one of those.